Implementando produtor de primos que marca os multiplos

+++ Ep1/src/ep1/calculadores/CalculadorAlgoritmo2.java
@@ -32,7 +32,7 @@ public class CalculadorAlgoritmo2 extends Thread {
 	}
 
 	private BigInteger primo(int i) {
-		return lista.get(i);
+		return BigInteger.valueOf(lista.get(i));
 	}
 
 	public BigInteger getSoma() {
